+++ frontend/package.json
@@ -44,5 +44,6 @@
   },
   "devDependencies": {
     "prettier": "2.4.1"
-  }
+  },
+  "proxy" : "http://localhost:8080"
 }
