+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import "./App.css"
-import { Route, Switch } from "react-router"
+import {Route, Switch, useHistory} from "react-router"
 import styled from "styled-components/macro"
 import Header from "./components/header/Header"
 import LoginPage from "./components/pages/LoginPage"
@@ -7,8 +7,33 @@ import Homepage from "./components/pages/homepage/Homepage"
 import RecommendationForm from "./components/pages/recommandationForm/RecommendationForm"
 import Footer from "./components/footer/Footer"
 import Recommendations from "./components/pages/recommandations/Recommendations";
+import {useState} from "react";
+import {getRecommendations} from "./utils/apiService/WineAPIService";
 
 function App() {
+    const [wineConfiguration, setWineConfiguration] = useState({
+        occasion: "",
+        wineStyle: "",
+        region: "",
+        alcohol: "",
+        taste: "",
+    })
+    const [recommendations,setRecommendations]=useState([])
+
+    const history=useHistory()
+
+    const handleSubmit = () => {
+        getRecommendations(wineConfiguration)
+            .then(result =>{
+                setRecommendations(result)
+            })
+        history.push("/recommendations")
+    }
+
+    const handleDataInput = (name, value) => {
+        setWineConfiguration({ ...wineConfiguration, [name]: value })
+    }
+
    return (
       <PageLayout>
          <Header />
@@ -20,10 +45,10 @@ function App() {
                <Homepage />
             </Route>
             <Route path="/wineConfiguration">
-               <RecommendationForm />
+               <RecommendationForm handleSubmit={handleSubmit} handleDataInput={handleDataInput} wineConfiguration={wineConfiguration}/>
             </Route>
              <Route path="/recommendations">
-                 <Recommendations/>
+                 <Recommendations recommendations={recommendations}/>
              </Route>
          </Switch>
          <Footer />
